+++ src/index.ts
@@ -1,0 +1,3 @@
+import createJetvilCMS from "./lib/cm-api/config/createJetvilCMS";
+
+export default () => createJetvilCMS();