+++ src/lib/cm-api/routes/schemaRouter.ts
@@ -1,0 +1,93 @@
+/* eslint-disable no-console */
+import isFalsy from "@jetvil/types/dist/lib/validate/types/isFalsy";
+import isTruthyExtended from "@jetvil/types/dist/lib/validate/types/isTruthyExtended";
+import { PrismaClient } from "@prisma/client";
+import { Router } from "express";
+import { logger } from "../config/logger";
+import { getPrismaFilter, getPrismaMethod } from "../data/prismaUtils";
+import { actions, functionType, methodTypes, middlewareType } from "./types";
+
+const createSchemaRouter = ({
+  client,
+  schema,
+  methods,
+  middleware,
+  verbose,
+}: {
+  client: PrismaClient;
+  schema: string;
+  methods: Array<methodTypes>;
+  middleware: middlewareType;
+  verbose: boolean;
+}) => {
+  logger.info(`Creating router for schema:: ${schema}...`, verbose);
+  const router: Router = Router();
+  const routes: Array<{ method: methodTypes; url: string; handler: functionType }> = [];
+
+  logger.info(`Creating routes for schema:: ${schema}...`, verbose);
+  methods.forEach((method: methodTypes) => {
+    const route = {
+      method: method as methodTypes,
+      url: "/" + schema + actions[method as methodTypes],
+      handler: async (req: any, res: any) => {
+        try {
+          const prismaMethod = getPrismaMethod(method);
+          const func = client[schema][prismaMethod];
+          const filter = getPrismaFilter(method);
+
+          const data: object = isTruthyExtended(req.body)
+            ? req.body
+            : isTruthyExtended(req.params)
+            ? req.params
+            : isTruthyExtended(req.query)
+            ? req.query
+            : undefined;
+          if (isFalsy(data)) {
+            throw new Error("No data provided");
+          }
+          const result = await func({ [filter]: data });
+          return res.json(result);
+        } catch (error: any) {
+          logger.error(error, verbose);
+          res.status(500).json({ error: error.message });
+        }
+      },
+    };
+
+    logger.info(`Adding route for method:: ${method} on schema:: ${schema}...`, verbose);
+    routes.push(route);
+  });
+
+  routes.push({
+    method: "get",
+    url: `/${schema}s`,
+    handler: async (_req: any, _res: any) => {
+      // const EitherOr: "AND" | "OR" = _req.query.match || _req.query.filter;
+      try {
+        const filters = Object.keys(_req.query).map((key) => {
+          const splitted = key.split(":");
+          const field: string = splitted[0];
+          const operator: string = splitted[1];
+          return { [field]: { [operator]: _req.query[key] } };
+        });
+        const where = filters.reduce((acc, filter) => ({ ...acc, ...filter }), {});
+        const result = await client[schema].findMany({ where });
+        return _res.json(result);
+      } catch (error: any) {
+        logger.error(error, verbose);
+        return _res.status(500).json({ error: error.message });
+      }
+    },
+  });
+
+  logger.info(`Adding routes to router for schema:: ${schema}...`, verbose);
+  const schemaArray = middleware?.filter((m) => isFalsy(m?.schemas?.length) || m?.schemas?.includes(schema));
+  routes.forEach((route) => {
+    const methodArray = schemaArray?.filter((m) => isFalsy(m?.methods?.length) || m?.methods?.includes(route.method));
+
+    router[route.method](route.url, ...(methodArray.map((m) => m.handler) as any), route.handler);
+  });
+  return router;
+};
+
+export default createSchemaRouter;