+++ src/lib/index.ts
@@ -1,0 +1,3 @@
+import createJetvilCMS from "./cm-api/config/createJetvilCMS";
+
+export default () => createJetvilCMS();