+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # 0.1.0
 
-- Initial setup
+- Initial setup CM-api with prisma and express.